+++ R/rd.R
@@ -73,7 +73,7 @@ roclet_tags.roclet_rd <- function(x) {
     source = tag_markdown,
     template = tag_value,
     templateVar = tag_name_description,
-    title = tag_markdown_restricted,
+    title = tag_markdown,
     usage = tag_value
   )
 }